Schematic and PCD
This folder contains the schematics and the PCB designs. There are some mistakes with the PCB design they are as follows
There is no GSM module being used, instead it's an RTC so the whole SIM800L section can be discarded.
New connections should be made to the SDA and SDL pins.
The LCD doesn't have a pot so that should also be added

The PCB design has been done in KiCAD. I did it in linux.